backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Todo App"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlmodel==0.0.21",
    "psycopg[pool]==3.2.0",
    "pydantic>=2.12.3",
    "pydantic-settings==2.6.0",
    "python-jose==3.3.0",
    "bcrypt==4.2.0",
    "python-multipart==0.0.20",
    "gunicorn==23.0.0",
    "openai>=1.59.0",
    "openai-agents>=0.6.0",
    "email-validator>=2.0.0",
    "openrouter>=0.1.3",
    "alembic>=1.18.1",
    "mcp>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
